pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "medseg-lab"
version = "0.1.0"
description = "Starter lab for medical image segmentation experiments (UNETR baseline + CI-friendly synthetic tests)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "you@example.com" }]
dependencies = [
  "torch>=2.1",
  "pytorch-lightning>=2.2",
  "monai>=1.3",
  "numpy>=1.24",
  "pyyaml>=6.0",
  "tqdm>=4.66",
]

[project.optional-dependencies]
cpu = []
cuda = []
test = ["pytest>=7.0"]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["src"]
filterwarnings = [
  "ignore::UserWarning",
]
